add update logic to extra components

+++ pkg/bootstrap/steps/policy_agent.go
@@ -20,6 +20,16 @@ const (
 // NewInstallPolicyAgentStep creates the policy agent installation step
 func NewInstallPolicyAgentStep(config Config) BootstrapStep {
 	config.Logger.Warningf("please note that the Policy Agent requires cert-manager to be installed!")
+
+	if slices.Contains(config.ComponentsExtra.Existing, policyAgentController) {
+		config.Logger.Warningf("controller: %s is already installed", policyAgentController)
+		return BootstrapStep{
+			Name:  "existing policy agent installation",
+			Input: []StepInput{},
+			Step:  doNothingStep,
+		}
+	}
+
 	return BootstrapStep{
 		Name:  "install Policy Agent",
 		Input: []StepInput{},
@@ -29,11 +39,6 @@ func NewInstallPolicyAgentStep(config Config) BootstrapStep {
 
 // installPolicyAgent start installing policy agent helm chart
 func installPolicyAgent(input []StepInput, c *Config) ([]StepOutput, error) {
-	if slices.Contains(c.ComponentsExtra.Existing, policyAgentController) {
-		c.Logger.Warningf("policy agent is already installed!")
-		return []StepOutput{}, nil
-	}
-
 	c.Logger.Actionf(policyAgentInstallInfoMsg)
 
 	// download agent file
